Different math functions added, dynamically generating list of functions signatures and a few comments

+++ src/main/java/com/prysenko/MathParser/Parsers/Infix/Nodes/ValueNode.java
@@ -8,7 +8,7 @@ public class ValueNode extends AbstractNode {
     }
 
     @Override
-    public int eval() {
-        return (int)value;
+    public double eval() {
+        return value;
     }
 }
